+++ classes/BackyardMysqli.php
@@ -23,7 +23,7 @@ class BackyardMysqli extends \mysqli
 
     /**
      * \mysqli wrapper with logger
-     * Sets the connection charset to utf-8 and collation to utf8_general_ci
+     * Sets the connection charset to utf8 by default
      *
      * @param string $host_port accepts either hostname (or IPv4) or hostname:port
      * To open as persistent use: $connection = new backyard_mysqli('p:' . $dbhost, $dbuser, $dbpass, $dbname);
@@ -31,10 +31,11 @@ class BackyardMysqli extends \mysqli
      * @param string $pass password
      * @param string $db database name
      * @param BackyardError|null $logger PSR-3 logger
+     * @param string $charset MySQL connection charset
      *
      * @todo add IPv6 , e.g ::1 as $host_port
      */
-    public function __construct($host_port, $user, $pass, $db, ?BackyardError $logger = null)
+    public function __construct($host_port, $user, $pass, $db, ?BackyardError $logger = null, $charset = 'utf8')
     {
         //error_log("debug: " . __CLASS__ . ' ' . __METHOD__);
         $this->logger = is_null($logger) ? new NullLogger() : $logger;
@@ -73,12 +74,9 @@ class BackyardMysqli extends \mysqli
             );
         }
 
-        //change character set to utf8
-        // todo this condition should enforce at least some utf8, but it also enforces utf8 over utf8mb4.
-        // Is there a way to see what utf8 is the table already counting on and then enforce the proper utf8
-        // communication? Or would it be safer for backward compatibility to add it as another argument to constructor?
-        if (!$this->set_charset("utf8")) {
-            $this->logger->log(2, sprintf("Error loading character set utf8: %s\n", $this->error));
+        //change character set
+        if (!$this->set_charset($charset)) {
+            $this->logger->log(2, sprintf("Error loading character set %s: %s\n", $charset, $this->error));
         }
     }